fix int to size_it

+++ sources/quadtree.cpp
@@ -259,7 +259,7 @@ void QuadTree::updateParticles(double step) {
     
     // We compute the forces exerted on the particles
     // CAN BE OPTIMIZED BY OPENMP
-    for (int i = 0; i < particles->size(); ++i) {
+    for (size_t i = 0; i < particles->size(); ++i) {
         Particle* particle = (*particles)[i];
         for (QuadTree* node : poOfSubtree[*rankMPI]) {
             double fx = 0.0, fy = 0.0;
@@ -275,7 +275,7 @@ void QuadTree::updateParticles(double step) {
 
     // On met à jour les vitesses et positions des particules
     // CAN BE OPTIMIZED BY OPENMP
-    for (int i = 0; i < particles->size(); ++i) {
+    for (size_t i = 0; i < particles->size(); ++i) {
         Particle* particle = (*particles)[i];
 
         // We print the force exerted on the particle
